+++ litellm/llms/bedrock/chat/__init__.py
@@ -1,0 +1,2 @@
+from .converse_handler import BedrockConverseLLM
+from .invoke_handler import BedrockLLM
