+++ litellm/llms/bedrock/chat/invoke_handler.py
@@ -52,8 +52,8 @@ from litellm.types.llms.openai import (
 from litellm.types.utils import GenericStreamingChunk as GChunk
 from litellm.utils import CustomStreamWrapper, ModelResponse, Usage, get_secret
 
-from ..base_aws_llm import BaseAWSLLM
-from ..prompt_templates.factory import (
+from ...base_aws_llm import BaseAWSLLM
+from ...prompt_templates.factory import (
     _bedrock_converse_messages_pt,
     _bedrock_tools_pt,
     cohere_message_pt,
@@ -64,7 +64,8 @@ from ..prompt_templates.factory import (
     parse_xml_params,
     prompt_factory,
 )
-from .common_utils import BedrockError, ModelResponseIterator, get_runtime_endpoint
+from ..common_utils import BedrockError, ModelResponseIterator, get_bedrock_tool_name
+from .converse_transformation import AmazonConverseConfig
 
 BEDROCK_CONVERSE_MODELS = [
     "anthropic.claude-3-5-sonnet-20240620-v1:0",
@@ -225,10 +226,9 @@ async def make_call(
             raise BedrockError(status_code=response.status_code, message=response.text)
 
         if "ai21" in api_base:
-            aws_bedrock_process_response = BedrockConverseLLM()
             model_response: (
                 ModelResponse
-            ) = aws_bedrock_process_response.process_response(
+            ) = litellm.AmazonConverseConfig()._transform_response(
                 model=model,
                 response=response,
                 model_response=litellm.ModelResponse(),
@@ -264,59 +264,6 @@ async def make_call(
         raise BedrockError(status_code=408, message="Timeout error occurred.")
     except Exception as e:
         raise BedrockError(status_code=500, message=str(e))
-
-
-def make_sync_call(
-    client: Optional[HTTPHandler],
-    api_base: str,
-    headers: dict,
-    data: str,
-    model: str,
-    messages: list,
-    logging_obj,
-):
-    if client is None:
-        client = _get_httpx_client()  # Create a new client if none provided
-
-    response = client.post(
-        api_base,
-        headers=headers,
-        data=data,
-        stream=True if "ai21" not in api_base else False,
-    )
-
-    if response.status_code != 200:
-        raise BedrockError(status_code=response.status_code, message=response.read())
-
-    if "ai21" in api_base:
-        aws_bedrock_process_response = BedrockConverseLLM()
-        model_response: ModelResponse = aws_bedrock_process_response.process_response(
-            model=model,
-            response=response,
-            model_response=litellm.ModelResponse(),
-            stream=True,
-            logging_obj=logging_obj,
-            optional_params={},
-            api_key="",
-            data=data,
-            messages=messages,
-            print_verbose=litellm.print_verbose,
-            encoding=litellm.encoding,
-        )  # type: ignore
-        completion_stream: Any = MockResponseIterator(model_response=model_response)
-    else:
-        decoder = AWSEventStreamDecoder(model=model)
-        completion_stream = decoder.iter_bytes(response.iter_bytes(chunk_size=1024))
-
-    # LOGGING
-    logging_obj.post_call(
-        input=messages,
-        api_key="",
-        original_response="first stream response received",
-        additional_args={"complete_input_dict": data},
-    )
-
-    return completion_stream
 
 
 class BedrockLLM(BaseAWSLLM):
@@ -417,6 +364,7 @@ class BedrockLLM(BaseAWSLLM):
         except:
             raise BedrockError(message=response.text, status_code=422)
 
+        outputText: Optional[str] = None
         try:
             if provider == "cohere":
                 if "text" in completion_response:
@@ -566,23 +514,27 @@ class BedrockLLM(BaseAWSLLM):
 
         try:
             if (
-                len(outputText) > 0
+                outputText is not None
+                and len(outputText) > 0
                 and hasattr(model_response.choices[0], "message")
-                and getattr(model_response.choices[0].message, "tool_calls", None)
+                and getattr(model_response.choices[0].message, "tool_calls", None)  # type: ignore
                 is None
             ):
-                model_response.choices[0].message.content = outputText
+                model_response.choices[0].message.content = outputText  # type: ignore
             elif (
                 hasattr(model_response.choices[0], "message")
-                and getattr(model_response.choices[0].message, "tool_calls", None)
+                and getattr(model_response.choices[0].message, "tool_calls", None)  # type: ignore
                 is not None
             ):
                 pass
             else:
                 raise Exception()
-        except:
+        except Exception as e:
             raise BedrockError(
-                message=json.dumps(outputText), status_code=response.status_code
+                message="Error parsing received text={}.\nError-{}".format(
+                    outputText, str(e)
+                ),
+                status_code=response.status_code,
             )
 
         if stream and provider == "ai21":
@@ -594,8 +546,8 @@ class BedrockLLM(BaseAWSLLM):
             streaming_choice = litellm.utils.StreamingChoices()
             streaming_choice.index = model_response.choices[0].index
             delta_obj = litellm.utils.Delta(
-                content=getattr(model_response.choices[0].message, "content", None),
-                role=model_response.choices[0].message.role,
+                content=getattr(model_response.choices[0].message, "content", None),  # type: ignore
+                role=model_response.choices[0].message.role,  # type: ignore
             )
             streaming_choice.delta = delta_obj
             streaming_model_response.choices = [streaming_choice]
@@ -731,7 +683,7 @@ class BedrockLLM(BaseAWSLLM):
         )
 
         ### SET RUNTIME ENDPOINT ###
-        endpoint_url, proxy_endpoint_url = get_runtime_endpoint(
+        endpoint_url, proxy_endpoint_url = self.get_runtime_endpoint(
             api_base=api_base,
             aws_bedrock_runtime_endpoint=aws_bedrock_runtime_endpoint,
             aws_region_name=aws_region_name,
@@ -1002,7 +954,7 @@ class BedrockLLM(BaseAWSLLM):
             response.raise_for_status()
         except httpx.HTTPStatusError as err:
             error_code = err.response.status_code
-            raise BedrockError(status_code=error_code, message=response.text)
+            raise BedrockError(status_code=error_code, message=err.response.text)
         except httpx.TimeoutException as e:
             raise BedrockError(status_code=408, message="Timeout error occurred.")
 
@@ -1113,725 +1065,6 @@ class BedrockLLM(BaseAWSLLM):
         return super().embedding(*args, **kwargs)
 
 
-class AmazonConverseConfig:
-    """
-    Reference - https://docs.aws.amazon.com/bedrock/latest/APIReference/API_runtime_Converse.html
-    #2 - https://docs.aws.amazon.com/bedrock/latest/userguide/conversation-inference.html#conversation-inference-supported-models-features
-    """
-
-    maxTokens: Optional[int]
-    stopSequences: Optional[List[str]]
-    temperature: Optional[int]
-    topP: Optional[int]
-
-    def __init__(
-        self,
-        maxTokens: Optional[int] = None,
-        stopSequences: Optional[List[str]] = None,
-        temperature: Optional[int] = None,
-        topP: Optional[int] = None,
-    ) -> None:
-        locals_ = locals()
-        for key, value in locals_.items():
-            if key != "self" and value is not None:
-                setattr(self.__class__, key, value)
-
-    @classmethod
-    def get_config(cls):
-        return {
-            k: v
-            for k, v in cls.__dict__.items()
-            if not k.startswith("__")
-            and not isinstance(
-                v,
-                (
-                    types.FunctionType,
-                    types.BuiltinFunctionType,
-                    classmethod,
-                    staticmethod,
-                ),
-            )
-            and v is not None
-        }
-
-    def get_supported_openai_params(self, model: str) -> List[str]:
-        supported_params = [
-            "max_tokens",
-            "max_completion_tokens",
-            "stream",
-            "stream_options",
-            "stop",
-            "temperature",
-            "top_p",
-            "extra_headers",
-            "response_format",
-        ]
-
-        if (
-            model.startswith("anthropic")
-            or model.startswith("mistral")
-            or model.startswith("cohere")
-            or model.startswith("meta.llama3-1")
-        ):
-            supported_params.append("tools")
-
-        if model.startswith("anthropic") or model.startswith("mistral"):
-            # only anthropic and mistral support tool choice config. otherwise (E.g. cohere) will fail the call - https://docs.aws.amazon.com/bedrock/latest/APIReference/API_runtime_ToolChoice.html
-            supported_params.append("tool_choice")
-
-        return supported_params
-
-    def map_tool_choice_values(
-        self, model: str, tool_choice: Union[str, dict], drop_params: bool
-    ) -> Optional[ToolChoiceValuesBlock]:
-        if tool_choice == "none":
-            if litellm.drop_params is True or drop_params is True:
-                return None
-            else:
-                raise litellm.utils.UnsupportedParamsError(
-                    message="Bedrock doesn't support tool_choice={}. To drop it from the call, set `litellm.drop_params = True.".format(
-                        tool_choice
-                    ),
-                    status_code=400,
-                )
-        elif tool_choice == "required":
-            return ToolChoiceValuesBlock(any={})
-        elif tool_choice == "auto":
-            return ToolChoiceValuesBlock(auto={})
-        elif isinstance(tool_choice, dict):
-            # only supported for anthropic + mistral models - https://docs.aws.amazon.com/bedrock/latest/APIReference/API_runtime_ToolChoice.html
-            specific_tool = SpecificToolChoiceBlock(
-                name=tool_choice.get("function", {}).get("name", "")
-            )
-            return ToolChoiceValuesBlock(tool=specific_tool)
-        else:
-            raise litellm.utils.UnsupportedParamsError(
-                message="Bedrock doesn't support tool_choice={}. Supported tool_choice values=['auto', 'required', json object]. To drop it from the call, set `litellm.drop_params = True.".format(
-                    tool_choice
-                ),
-                status_code=400,
-            )
-
-    def get_supported_image_types(self) -> List[str]:
-        return ["png", "jpeg", "gif", "webp"]
-
-    def map_openai_params(
-        self,
-        model: str,
-        non_default_params: dict,
-        optional_params: dict,
-        drop_params: bool,
-    ) -> dict:
-        for param, value in non_default_params.items():
-            if param == "response_format":
-                json_schema: Optional[dict] = None
-                schema_name: str = ""
-                if "response_schema" in value:
-                    json_schema = value["response_schema"]
-                    schema_name = "json_tool_call"
-                elif "json_schema" in value:
-                    json_schema = value["json_schema"]["schema"]
-                    schema_name = value["json_schema"]["name"]
-                """
-                Follow similar approach to anthropic - translate to a single tool call. 
-
-                When using tools in this way: - https://docs.anthropic.com/en/docs/build-with-claude/tool-use#json-mode
-                - You usually want to provide a single tool
-                - You should set tool_choice (see Forcing tool use) to instruct the model to explicitly use that tool
-                - Remember that the model will pass the input to the tool, so the name of the tool and description should be from the model’s perspective.
-                """
-                if json_schema is not None:
-                    _tool_choice = self.map_tool_choice_values(
-                        model=model, tool_choice="required", drop_params=drop_params  # type: ignore
-                    )
-
-                    _tool = ChatCompletionToolParam(
-                        type="function",
-                        function=ChatCompletionToolParamFunctionChunk(
-                            name=schema_name, parameters=json_schema
-                        ),
-                    )
-
-                    optional_params["tools"] = [_tool]
-                    optional_params["tool_choice"] = _tool_choice
-                    optional_params["json_mode"] = True
-                else:
-                    if litellm.drop_params is True or drop_params is True:
-                        pass
-                    else:
-                        raise litellm.utils.UnsupportedParamsError(
-                            message="Bedrock doesn't support response_format={}. To drop it from the call, set `litellm.drop_params = True.".format(
-                                value
-                            ),
-                            status_code=400,
-                        )
-            if param == "max_tokens" or param == "max_completion_tokens":
-                optional_params["maxTokens"] = value
-            if param == "stream":
-                optional_params["stream"] = value
-            if param == "stop":
-                if isinstance(value, str):
-                    if len(value) == 0:  # converse raises error for empty strings
-                        continue
-                    value = [value]
-                optional_params["stopSequences"] = value
-            if param == "temperature":
-                optional_params["temperature"] = value
-            if param == "top_p":
-                optional_params["topP"] = value
-            if param == "tools":
-                optional_params["tools"] = value
-            if param == "tool_choice":
-                _tool_choice_value = self.map_tool_choice_values(
-                    model=model, tool_choice=value, drop_params=drop_params  # type: ignore
-                )
-                if _tool_choice_value is not None:
-                    optional_params["tool_choice"] = _tool_choice_value
-        return optional_params
-
-
-class BedrockConverseLLM(BaseAWSLLM):
-    def __init__(self) -> None:
-        super().__init__()
-
-    def process_response(
-        self,
-        model: str,
-        response: Union[requests.Response, httpx.Response],
-        model_response: ModelResponse,
-        stream: bool,
-        logging_obj: Optional[Logging],
-        optional_params: dict,
-        api_key: str,
-        data: Union[dict, str],
-        messages: List,
-        print_verbose,
-        encoding,
-    ) -> Union[ModelResponse, CustomStreamWrapper]:
-
-        ## LOGGING
-        if logging_obj is not None:
-            logging_obj.post_call(
-                input=messages,
-                api_key=api_key,
-                original_response=response.text,
-                additional_args={"complete_input_dict": data},
-            )
-        print_verbose(f"raw model_response: {response.text}")
-        json_mode: Optional[bool] = optional_params.pop("json_mode", None)
-        ## RESPONSE OBJECT
-        try:
-            completion_response = ConverseResponseBlock(**response.json())  # type: ignore
-        except Exception as e:
-            raise BedrockError(
-                message="Received={}, Error converting to valid response block={}. File an issue if litellm error - https://github.com/BerriAI/litellm/issues".format(
-                    response.text, str(e)
-                ),
-                status_code=422,
-            )
-
-        """
-        Bedrock Response Object has optional message block 
-
-        completion_response["output"].get("message", None)
-
-        A message block looks like this (Example 1): 
-        "output": {
-            "message": {
-                "role": "assistant",
-                "content": [
-                    {
-                        "text": "Is there anything else you'd like to talk about? Perhaps I can help with some economic questions or provide some information about economic concepts?"
-                    }
-                ]
-            }
-        },
-        (Example 2):
-        "output": {
-            "message": {
-                "role": "assistant",
-                "content": [
-                    {
-                        "toolUse": {
-                            "toolUseId": "tooluse_hbTgdi0CSLq_hM4P8csZJA",
-                            "name": "top_song",
-                            "input": {
-                                "sign": "WZPZ"
-                            }
-                        }
-                    }
-                ]
-            }
-        }
-
-        """
-        message: Optional[MessageBlock] = completion_response["output"]["message"]
-        chat_completion_message: ChatCompletionResponseMessage = {"role": "assistant"}
-        content_str = ""
-        tools: List[ChatCompletionToolCallChunk] = []
-        if message is not None:
-            for idx, content in enumerate(message["content"]):
-                """
-                - Content is either a tool response or text
-                """
-                if "text" in content:
-                    content_str += content["text"]
-                if "toolUse" in content:
-
-                    ## check tool name was formatted by litellm
-                    _response_tool_name = content["toolUse"]["name"]
-                    response_tool_name = get_bedrock_tool_name(
-                        response_tool_name=_response_tool_name
-                    )
-                    _function_chunk = ChatCompletionToolCallFunctionChunk(
-                        name=response_tool_name,
-                        arguments=json.dumps(content["toolUse"]["input"]),
-                    )
-
-                    _tool_response_chunk = ChatCompletionToolCallChunk(
-                        id=content["toolUse"]["toolUseId"],
-                        type="function",
-                        function=_function_chunk,
-                        index=idx,
-                    )
-                    tools.append(_tool_response_chunk)
-        chat_completion_message["content"] = content_str
-
-        if json_mode is True and tools is not None and len(tools) == 1:
-            # to support 'json_schema' logic on bedrock models
-            json_mode_content_str: Optional[str] = tools[0]["function"].get("arguments")
-            if json_mode_content_str is not None:
-                chat_completion_message["content"] = json_mode_content_str
-        else:
-            chat_completion_message["tool_calls"] = tools
-
-        ## CALCULATING USAGE - bedrock returns usage in the headers
-        input_tokens = completion_response["usage"]["inputTokens"]
-        output_tokens = completion_response["usage"]["outputTokens"]
-        total_tokens = completion_response["usage"]["totalTokens"]
-
-        model_response.choices = [
-            litellm.Choices(
-                finish_reason=map_finish_reason(completion_response["stopReason"]),
-                index=0,
-                message=litellm.Message(**chat_completion_message),
-            )
-        ]
-        model_response.created = int(time.time())
-        model_response.model = model
-        usage = Usage(
-            prompt_tokens=input_tokens,
-            completion_tokens=output_tokens,
-            total_tokens=total_tokens,
-        )
-        setattr(model_response, "usage", usage)
-
-        # Add "trace" from Bedrock guardrails - if user has opted in to returning it
-        if "trace" in completion_response:
-            setattr(model_response, "trace", completion_response["trace"])
-
-        return model_response
-
-    def encode_model_id(self, model_id: str) -> str:
-        """
-        Double encode the model ID to ensure it matches the expected double-encoded format.
-        Args:
-            model_id (str): The model ID to encode.
-        Returns:
-            str: The double-encoded model ID.
-        """
-        return urllib.parse.quote(model_id, safe="")
-
-    async def async_streaming(
-        self,
-        model: str,
-        messages: list,
-        api_base: str,
-        model_response: ModelResponse,
-        print_verbose: Callable,
-        data: str,
-        timeout: Optional[Union[float, httpx.Timeout]],
-        encoding,
-        logging_obj,
-        stream,
-        optional_params: dict,
-        litellm_params=None,
-        logger_fn=None,
-        headers={},
-        client: Optional[AsyncHTTPHandler] = None,
-    ) -> CustomStreamWrapper:
-        streaming_response = CustomStreamWrapper(
-            completion_stream=None,
-            make_call=partial(
-                make_call,
-                client=client,
-                api_base=api_base,
-                headers=headers,
-                data=data,
-                model=model,
-                messages=messages,
-                logging_obj=logging_obj,
-            ),
-            model=model,
-            custom_llm_provider="bedrock",
-            logging_obj=logging_obj,
-        )
-        return streaming_response
-
-    async def async_completion(
-        self,
-        model: str,
-        messages: list,
-        api_base: str,
-        model_response: ModelResponse,
-        print_verbose: Callable,
-        data: str,
-        timeout: Optional[Union[float, httpx.Timeout]],
-        encoding,
-        logging_obj,
-        stream,
-        optional_params: dict,
-        litellm_params=None,
-        logger_fn=None,
-        headers={},
-        client: Optional[AsyncHTTPHandler] = None,
-    ) -> Union[ModelResponse, CustomStreamWrapper]:
-        if client is None or not isinstance(client, AsyncHTTPHandler):
-            _params = {}
-            if timeout is not None:
-                if isinstance(timeout, float) or isinstance(timeout, int):
-                    timeout = httpx.Timeout(timeout)
-                _params["timeout"] = timeout
-            client = get_async_httpx_client(
-                params=_params, llm_provider=litellm.LlmProviders.BEDROCK
-            )
-        else:
-            client = client  # type: ignore
-
-        try:
-            response = await client.post(url=api_base, headers=headers, data=data)  # type: ignore
-            response.raise_for_status()
-        except httpx.HTTPStatusError as err:
-            error_code = err.response.status_code
-            raise BedrockError(status_code=error_code, message=err.response.text)
-        except httpx.TimeoutException as e:
-            raise BedrockError(status_code=408, message="Timeout error occurred.")
-
-        return self.process_response(
-            model=model,
-            response=response,
-            model_response=model_response,
-            stream=stream if isinstance(stream, bool) else False,
-            logging_obj=logging_obj,
-            api_key="",
-            data=data,
-            messages=messages,
-            print_verbose=print_verbose,
-            optional_params=optional_params,
-            encoding=encoding,
-        )
-
-    def completion(
-        self,
-        model: str,
-        messages: list,
-        api_base: Optional[str],
-        custom_prompt_dict: dict,
-        model_response: ModelResponse,
-        print_verbose: Callable,
-        encoding,
-        logging_obj,
-        optional_params: dict,
-        acompletion: bool,
-        timeout: Optional[Union[float, httpx.Timeout]],
-        litellm_params: dict,
-        logger_fn=None,
-        extra_headers: Optional[dict] = None,
-        client: Optional[Union[AsyncHTTPHandler, HTTPHandler]] = None,
-    ):
-        try:
-            import boto3
-            from botocore.auth import SigV4Auth
-            from botocore.awsrequest import AWSRequest
-            from botocore.credentials import Credentials
-        except ImportError:
-            raise ImportError("Missing boto3 to call bedrock. Run 'pip install boto3'.")
-
-        ## SETUP ##
-        stream = optional_params.pop("stream", None)
-        modelId = optional_params.pop("model_id", None)
-        if modelId is not None:
-            modelId = self.encode_model_id(model_id=modelId)
-        else:
-            modelId = model
-
-        provider = model.split(".")[0]
-
-        ## CREDENTIALS ##
-        # pop aws_secret_access_key, aws_access_key_id, aws_region_name from kwargs, since completion calls fail with them
-        aws_secret_access_key = optional_params.pop("aws_secret_access_key", None)
-        aws_access_key_id = optional_params.pop("aws_access_key_id", None)
-        aws_session_token = optional_params.pop("aws_session_token", None)
-        aws_region_name = optional_params.pop("aws_region_name", None)
-        aws_role_name = optional_params.pop("aws_role_name", None)
-        aws_session_name = optional_params.pop("aws_session_name", None)
-        aws_profile_name = optional_params.pop("aws_profile_name", None)
-        aws_bedrock_runtime_endpoint = optional_params.pop(
-            "aws_bedrock_runtime_endpoint", None
-        )  # https://bedrock-runtime.{region_name}.amazonaws.com
-        aws_web_identity_token = optional_params.pop("aws_web_identity_token", None)
-        aws_sts_endpoint = optional_params.pop("aws_sts_endpoint", None)
-
-        ### SET REGION NAME ###
-        if aws_region_name is None:
-            # check env #
-            litellm_aws_region_name = get_secret("AWS_REGION_NAME", None)
-
-            if litellm_aws_region_name is not None and isinstance(
-                litellm_aws_region_name, str
-            ):
-                aws_region_name = litellm_aws_region_name
-
-            standard_aws_region_name = get_secret("AWS_REGION", None)
-            if standard_aws_region_name is not None and isinstance(
-                standard_aws_region_name, str
-            ):
-                aws_region_name = standard_aws_region_name
-
-            if aws_region_name is None:
-                aws_region_name = "us-west-2"
-
-        credentials: Credentials = self.get_credentials(
-            aws_access_key_id=aws_access_key_id,
-            aws_secret_access_key=aws_secret_access_key,
-            aws_session_token=aws_session_token,
-            aws_region_name=aws_region_name,
-            aws_session_name=aws_session_name,
-            aws_profile_name=aws_profile_name,
-            aws_role_name=aws_role_name,
-            aws_web_identity_token=aws_web_identity_token,
-            aws_sts_endpoint=aws_sts_endpoint,
-        )
-
-        ### SET RUNTIME ENDPOINT ###
-        endpoint_url, proxy_endpoint_url = get_runtime_endpoint(
-            api_base=api_base,
-            aws_bedrock_runtime_endpoint=aws_bedrock_runtime_endpoint,
-            aws_region_name=aws_region_name,
-        )
-        if (stream is not None and stream is True) and provider != "ai21":
-            endpoint_url = f"{endpoint_url}/model/{modelId}/converse-stream"
-            proxy_endpoint_url = f"{proxy_endpoint_url}/model/{modelId}/converse-stream"
-        else:
-            endpoint_url = f"{endpoint_url}/model/{modelId}/converse"
-            proxy_endpoint_url = f"{proxy_endpoint_url}/model/{modelId}/converse"
-
-        sigv4 = SigV4Auth(credentials, "bedrock", aws_region_name)
-
-        # Separate system prompt from rest of message
-        system_prompt_indices = []
-        system_content_blocks: List[SystemContentBlock] = []
-        for idx, message in enumerate(messages):
-            if message["role"] == "system":
-                _system_content_block: Optional[SystemContentBlock] = None
-                if isinstance(message["content"], str) and len(message["content"]) > 0:
-                    _system_content_block = SystemContentBlock(text=message["content"])
-                elif isinstance(message["content"], list):
-                    for m in message["content"]:
-                        if m.get("type", "") == "text" and len(m["text"]) > 0:
-                            _system_content_block = SystemContentBlock(text=m["text"])
-                if _system_content_block is not None:
-                    system_content_blocks.append(_system_content_block)
-                system_prompt_indices.append(idx)
-        if len(system_prompt_indices) > 0:
-            for idx in reversed(system_prompt_indices):
-                messages.pop(idx)
-
-        inference_params = copy.deepcopy(optional_params)
-        additional_request_keys = []
-        additional_request_params = {}
-        supported_converse_params = AmazonConverseConfig.__annotations__.keys()
-        supported_tool_call_params = ["tools", "tool_choice"]
-        supported_guardrail_params = ["guardrailConfig"]
-        json_mode: Optional[bool] = inference_params.pop(
-            "json_mode", None
-        )  # used for handling json_schema
-        ## TRANSFORMATION ##
-
-        bedrock_messages: List[MessageBlock] = _bedrock_converse_messages_pt(
-            messages=messages,
-            model=model,
-            llm_provider="bedrock_converse",
-            user_continue_message=litellm_params.pop("user_continue_message", None),
-        )
-
-        # send all model-specific params in 'additional_request_params'
-        for k, v in inference_params.items():
-            if (
-                k not in supported_converse_params
-                and k not in supported_tool_call_params
-                and k not in supported_guardrail_params
-            ):
-                additional_request_params[k] = v
-                additional_request_keys.append(k)
-        for key in additional_request_keys:
-            inference_params.pop(key, None)
-
-        bedrock_tools: List[ToolBlock] = _bedrock_tools_pt(
-            inference_params.pop("tools", [])
-        )
-        bedrock_tool_config: Optional[ToolConfigBlock] = None
-        if len(bedrock_tools) > 0:
-            tool_choice_values: ToolChoiceValuesBlock = inference_params.pop(
-                "tool_choice", None
-            )
-            bedrock_tool_config = ToolConfigBlock(
-                tools=bedrock_tools,
-            )
-            if tool_choice_values is not None:
-                bedrock_tool_config["toolChoice"] = tool_choice_values
-
-        _data: RequestObject = {
-            "messages": bedrock_messages,
-            "additionalModelRequestFields": additional_request_params,
-            "system": system_content_blocks,
-            "inferenceConfig": InferenceConfig(**inference_params),
-        }
-
-        # Guardrail Config
-        guardrail_config: Optional[GuardrailConfigBlock] = None
-        request_guardrails_config = inference_params.pop("guardrailConfig", None)
-        if request_guardrails_config is not None:
-            guardrail_config = GuardrailConfigBlock(**request_guardrails_config)
-            _data["guardrailConfig"] = guardrail_config
-
-        # Tool Config
-        if bedrock_tool_config is not None:
-            _data["toolConfig"] = bedrock_tool_config
-
-        data = json.dumps(_data)
-        ## COMPLETION CALL
-
-        headers = {"Content-Type": "application/json"}
-        if extra_headers is not None:
-            headers = {"Content-Type": "application/json", **extra_headers}
-        request = AWSRequest(
-            method="POST", url=endpoint_url, data=data, headers=headers
-        )
-        sigv4.add_auth(request)
-        if (
-            extra_headers is not None and "Authorization" in extra_headers
-        ):  # prevent sigv4 from overwriting the auth header
-            request.headers["Authorization"] = extra_headers["Authorization"]
-        prepped = request.prepare()
-
-        ## LOGGING
-        logging_obj.pre_call(
-            input=messages,
-            api_key="",
-            additional_args={
-                "complete_input_dict": data,
-                "api_base": proxy_endpoint_url,
-                "headers": prepped.headers,
-            },
-        )
-
-        ### ROUTING (ASYNC, STREAMING, SYNC)
-        if acompletion:
-            if isinstance(client, HTTPHandler):
-                client = None
-            if stream is True:
-                return self.async_streaming(
-                    model=model,
-                    messages=messages,
-                    data=data,
-                    api_base=proxy_endpoint_url,
-                    model_response=model_response,
-                    print_verbose=print_verbose,
-                    encoding=encoding,
-                    logging_obj=logging_obj,
-                    optional_params=optional_params,
-                    stream=True,
-                    litellm_params=litellm_params,
-                    logger_fn=logger_fn,
-                    headers=prepped.headers,
-                    timeout=timeout,
-                    client=client,
-                )  # type: ignore
-            ### ASYNC COMPLETION
-            return self.async_completion(
-                model=model,
-                messages=messages,
-                data=data,
-                api_base=proxy_endpoint_url,
-                model_response=model_response,
-                print_verbose=print_verbose,
-                encoding=encoding,
-                logging_obj=logging_obj,
-                optional_params=optional_params,
-                stream=stream,  # type: ignore
-                litellm_params=litellm_params,
-                logger_fn=logger_fn,
-                headers=prepped.headers,
-                timeout=timeout,
-                client=client,
-            )  # type: ignore
-
-        if stream is not None and stream is True:
-
-            streaming_response = CustomStreamWrapper(
-                completion_stream=None,
-                make_call=partial(
-                    make_sync_call,
-                    client=None,
-                    api_base=proxy_endpoint_url,
-                    headers=prepped.headers,  # type: ignore
-                    data=data,
-                    model=model,
-                    messages=messages,
-                    logging_obj=logging_obj,
-                ),
-                model=model,
-                custom_llm_provider="bedrock",
-                logging_obj=logging_obj,
-            )
-
-            return streaming_response
-        ### COMPLETION
-
-        if client is None or isinstance(client, AsyncHTTPHandler):
-            _params = {}
-            if timeout is not None:
-                if isinstance(timeout, float) or isinstance(timeout, int):
-                    timeout = httpx.Timeout(timeout)
-                _params["timeout"] = timeout
-            client = _get_httpx_client(_params)  # type: ignore
-        else:
-            client = client
-        try:
-            response = client.post(url=proxy_endpoint_url, headers=prepped.headers, data=data)  # type: ignore
-            response.raise_for_status()
-        except httpx.HTTPStatusError as err:
-            error_code = err.response.status_code
-            raise BedrockError(status_code=error_code, message=response.text)
-        except httpx.TimeoutException:
-            raise BedrockError(status_code=408, message="Timeout error occurred.")
-
-        return self.process_response(
-            model=model,
-            response=response,
-            model_response=model_response,
-            stream=stream if isinstance(stream, bool) else False,
-            logging_obj=logging_obj,
-            optional_params=optional_params,
-            api_key="",
-            data=data,
-            messages=messages,
-            print_verbose=print_verbose,
-            encoding=encoding,
-        )
-
-
 def get_response_stream_shape():
     global _response_stream_shape_cache
     if _response_stream_shape_cache is None:
@@ -1845,24 +1078,6 @@ def get_response_stream_shape():
         _response_stream_shape_cache = bedrock_service_model.shape_for("ResponseStream")
 
     return _response_stream_shape_cache
-
-
-def get_bedrock_tool_name(response_tool_name: str) -> str:
-    """
-    If litellm formatted the input tool name, we need to convert it back to the original name.
-
-    Args:
-        response_tool_name (str): The name of the tool as received from the response.
-
-    Returns:
-        str: The original name of the tool.
-    """
-
-    if response_tool_name in litellm.bedrock_tool_name_mappings.cache_dict:
-        response_tool_name = litellm.bedrock_tool_name_mappings.cache_dict[
-            response_tool_name
-        ]
-    return response_tool_name
 
 
 class AWSEventStreamDecoder:
