+++ litellm/main.py
@@ -2385,6 +2385,7 @@ def completion(
                     )
 
             if model in litellm.BEDROCK_CONVERSE_MODELS:
+
                 response = bedrock_converse_chat_completion.completion(
                     model=model,
                     messages=messages,
@@ -3570,7 +3571,7 @@ def embedding(
                 client=client,
                 timeout=timeout,
                 aembedding=aembedding,
-                litellm_params=litellm_params,
+                litellm_params={},
                 api_base=api_base,
                 print_verbose=print_verbose,
                 extra_headers=extra_headers,
