+++ litellm/proxy/_new_secret_config.yaml
@@ -1,7 +1,20 @@
 model_list:
-  - model_name: gpt-3.5-turbo
+  - model_name: fake-claude-endpoint
+    litellm_params:
+      model: anthropic.claude-3-sonnet-20240229-v1:0
+      api_base: https://exampleopenaiendpoint-production.up.railway.app
+      # aws_session_token: "IQoJb3JpZ2luX2VjELj//////////wEaCXVzLXdlc3QtMiJHMEUCIQDatCRVkIZERLcrR6P7Qd1vNfZ8r8xB/LUeaVaTW/lBTwIgAgmHSBe41d65GVRKSkpgVonjsCmOmAS7s/yklM9NsZcq3AEI4P//////////ARABGgw4ODg2MDIyMjM0MjgiDJrio0/CHYEfyt5EqyqwAfyWO4t3bFVWAOIwTyZ1N6lszeJKfMNus2hzVc+r73hia2Anv88uwPxNg2uqnXQNJumEo0DcBt30ZwOw03Isboy0d5l05h8gjb4nl9feyeKmKAnRdcqElrEWtCC1Qcefv78jQv53AbUipH1ssa5NPvptqZZpZYDPMlBEnV3YdvJJiuE23u2yOkCt+EoUJLaOYjOryoRyrSfbWB+JaUsB68R3rNTHzReeN3Nob/9Ic4HrMMmzmLcGOpgBZxclO4w8Z7i6TcVqbCwDOskxuR6bZaiFxKFG+9tDrWS7jaQKpq/YP9HUT0YwYpZplaBEEZR5sbIndg5yb4dRZrSHplblqKz8XLaUf5tuuyRJmwr96PTpw/dyEVk9gicFX6JfLBEv0v5rN2Z0JMFLdfIP4kC1U2PjcPOWoglWO3fLmJ4Lol2a3c5XDSMwMxjcJXq+c8Ue1v0="
+      aws_secret_access_key: os.environ/AWS_SECRET_ACCESS_KEY
+      aws_access_key_id: os.environ/AWS_ACCESS_KEY_ID
+  - model_name: gemini-vision
+    litellm_params:
+      model: vertex_ai/gemini-1.0-pro-vision-001
+      api_base: https://exampleopenaiendpoint-production.up.railway.app/v1/projects/adroit-crow-413218/locations/us-central1/publishers/google/models/gemini-1.0-pro-vision-001
+      vertex_project: "adroit-crow-413218"
+      vertex_location: "us-central1"
+  
+  - model_name: fake-openai-endpoint
     litellm_params:
       model: gpt-3.5-turbo
-
-router_settings:
-  model_group_alias: {"gpt-4": {"model": "gpt-3.5-turbo", "hidden": false}}
+      api_base: https://exampleopenaiendpoint-production.up.railway.app
+
