+++ litellm/tests/test_bedrock_completion.py
@@ -27,7 +27,7 @@ from litellm import (
     completion_cost,
     embedding,
 )
-from litellm.llms.bedrock.chat import BedrockLLM, ToolBlock
+from litellm.llms.bedrock.chat import BedrockLLM
 from litellm.llms.custom_httpx.http_handler import AsyncHTTPHandler, HTTPHandler
 from litellm.llms.prompt_templates.factory import _bedrock_tools_pt
 
@@ -1287,3 +1287,41 @@ def test_bedrock_converse_translation_tool_message():
             ],
         }
     ]
+
+
+def test_base_aws_llm_get_credentials():
+    import time
+
+    import boto3
+
+    from litellm.llms.base_aws_llm import BaseAWSLLM
+
+    start_time = time.time()
+    session = boto3.Session(
+        aws_access_key_id="test",
+        aws_secret_access_key="test2",
+        region_name="test3",
+    )
+    credentials = session.get_credentials().get_frozen_credentials()
+    end_time = time.time()
+
+    print(
+        "Total time for credentials - {}. Credentials - {}".format(
+            end_time - start_time, credentials
+        )
+    )
+
+    start_time = time.time()
+    credentials = BaseAWSLLM().get_credentials(
+        aws_access_key_id="test",
+        aws_secret_access_key="test2",
+        aws_region_name="test3",
+    )
+
+    end_time = time.time()
+
+    print(
+        "Total time for credentials - {}. Credentials - {}".format(
+            end_time - start_time, credentials.get_frozen_credentials()
+        )
+    )
